Move Rng trait to sub-module
